Struct hydro_cli::core::hydroflow_crate::HydroflowCrate

pub struct HydroflowCrate {
    src: PathBuf,
    on: Arc<RwLock<dyn Host>>,
    example: Option<String>,
    features: Option<Vec<String>>,
    outgoing_ports: HashMap<String, (Weak<RwLock<dyn Service>>, String)>,
    incoming_ports: HashMap<String, Arc<RwLock<dyn Host>>>,
    built_binary: Option<String>,
    launched_host: Option<Arc<dyn LaunchedHost>>,
    bound_pipes: HashMap<String, ConnectionPipe>,
    launched_binary: Option<Arc<RwLock<dyn LaunchedBinary>>>,
}

Fields
src: PathBufon: Arc<RwLock<dyn Host>>example: Option<String>features: Option<Vec<String>>outgoing_ports: HashMap<String, (Weak<RwLock<dyn Service>>, String)>

A mapping from output ports to the service that the port sends data to. Each dyn Service is guaranteed to be another HydroflowCrate.

incoming_ports: HashMap<String, Arc<RwLock<dyn Host>>>

A map of port names to the host that will be sending data to the port.

built_binary: Option<String>launched_host: Option<Arc<dyn LaunchedHost>>bound_pipes: HashMap<String, ConnectionPipe>

A map of port names to config for how other services can connect to this one. Only valid after ready has been called.

launched_binary: Option<Arc<RwLock<dyn LaunchedBinary>>>
Implementations

impl HydroflowCrate

pub fn new(
src: PathBuf,
on: Arc<RwLock<dyn Host>>,
example: Option<String>,
features: Option<Vec<String>>
) -> Self

pub fn add_incoming_port(&mut self, my_port: String, from: &HydroflowCrate)

pub fn add_outgoing_port(
&mut self,
my_port: String,
to: &Arc<RwLock<dyn Service>>,
to_port: String
)

pub async fn stdout(&self) -> Receiver<String>

pub async fn stderr(&self) -> Receiver<String>

pub async fn exit_code(&self) -> Option<i32>

fn build(&mut self) -> JoinHandle<String>