Trait hydro_cli::core::Service

pub trait Service: Send + Sync + Debug {
    fn collect_resources<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        resource_batch: &'life1 mut ResourceBatch
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn deploy<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        resource_result: &'life1 Arc<ResourceResult>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn ready<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn start<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn host(&self) -> Arc<RwLock<dyn Host>>;
    fn as_any(&self) -> &dyn Any;
    fn as_any_mut(&mut self) -> &mut dyn Any;
}

Required Methods

fn collect_resources<'life0, 'life1, 'async_trait>(
&'life0 mut self,
resource_batch: &'life1 mut ResourceBatch
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Makes requests for physical resources (servers) that this service needs to run. This should also perform any “free” computations (compilations) that are needed, because the deploy method will be called after these resources are allocated.

fn deploy<'life0, 'life1, 'async_trait>(
&'life0 mut self,
resource_result: &'life1 Arc<ResourceResult>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Connects to the acquired resources and prepares the service to be launched.

fn ready<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Launches the service, which should start listening for incoming network connections. The service should not start computing at this point.

fn start<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Starts the service by having it connect to other services and start computations.

fn host(&self) -> Arc<RwLock<dyn Host>>

fn as_any(&self) -> &dyn Any

fn as_any_mut(&mut self) -> &mut dyn Any